Moved output
Moved timings
