+++ crates/fta/src/main.rs
@@ -1,5 +1,7 @@
 use clap::Parser;
 use fta::analyze;
+use fta::output;
+use std::time::Instant;
 
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about = None)]
@@ -12,7 +14,15 @@ struct Cli {
 }
 
 pub fn main() {
+    // Start tracking execution time
+    let start = Instant::now();
+
     let cli = Cli::parse();
 
-    analyze(&cli.project, cli.format)
+    let findings = analyze(&cli.project);
+
+    // - Move to a metrics/perf module.
+    let elapsed = start.elapsed().as_secs_f64();
+
+    output::output(&findings, &cli.format, &elapsed);
 }
