+++ crates/fta/src/output/tests.rs
@@ -1,0 +1,6 @@
+#[cfg(test)]
+mod tests {
+    use crate::output::output;
+    use std::io::Write;
+    use tempfile::NamedTempFile;
+}
